# componenteA/src/main/java/uk/techiegt/clipe/componenteA/entities/Cliente.java
package uk.techiegt.clipe.componenteA.entities;

import jakarta.persistence.*;
import lombok.*;

@Entity @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
public class Cliente {
    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;
    private String nombre;
    private String correo;
}