# componenteA/src/main/java/uk/techiegt/clipe/componenteA/entities/Pedido.java
package uk.techiegt.clipe.componenteA.entities;

import jakarta.persistence.*;
import lombok.*;

@Entity @Getter @Setter @NoArgsConstructor @AllArgsConstructor @Builder
public class Pedido {
    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer id;

    private Integer clienteId;

    /** Total calculado usando la librería del componente C */
    private double total;
}