# componenteA/src/main/resources/ClienteApiSpecification.yaml
openapi: 3.0.1
info:
  title: API Componente A - Clientes y Pedidos
  version: 1.0.0
  description: API mínima para gestión de clientes y pedidos (Mysql)
servers:
  - url: http://localhost:8080
paths:
  /clientes:
    post:
      summary: Crear un cliente
      operationId: crearCliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClienteInput'
      responses:
        '201':
          description: Cliente creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '400':
          description: Datos inválidos
    get:
      summary: Listar todos los clientes
      operationId: listarClientes
      responses:
        '200':
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cliente'
  /clientes/{id}:
    get:
      summary: Obtener un cliente por ID
      operationId: obtenerClientePorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Cliente'
        '404':
          description: Cliente no encontrado
  /pedidos:
    post:
      summary: Crear un pedido
      operationId: crearPedido
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PedidoInput'
      responses:
        '201':
          description: Pedido creado correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '400':
          description: Datos inválidos
    get:
      summary: Listar todos los pedidos
      operationId: listarPedidos
      responses:
        '200':
          description: Lista de pedidos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Pedido'
  /pedidos/{id}:
    get:
      summary: Obtener un pedido por ID
      operationId: obtenerPedidoPorId
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Pedido encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pedido'
        '404':
          description: Pedido no encontrado

components:
  schemas:
    Cliente:
      type: object
      properties:
        id:
          type: integer
          format: int32
        nombre:
          type: string
        correo:
          type: string
    ClienteInput:
      type: object
      required:
        - nombre
        - correo
      properties:
        nombre:
          type: string
        correo:
          type: string
    Producto:
      type: object
      properties:
        nombre:
          type: string
        precio:
          type: number
          format: double
    Pedido:
      type: object
      properties:
        id:
          type: integer
        clienteId:
          type: integer
        productos:
          type: array
          items:
            $ref: '#/components/schemas/Producto'
        total:
          type: number
          format: double
    PedidoInput:
      type: object
      required:
        - clienteId
        - productos
      properties:
        clienteId:
          type: integer
        productos:
          type: array
          items:
            $ref: '#/components/schemas/Producto'